+++ internal/datastore/common/helpers.go
@@ -2,17 +2,12 @@ package common
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"strconv"
-	"strings"
 
-	"github.com/dustin/go-humanize"
 	"google.golang.org/protobuf/types/known/structpb"
 
 	"github.com/authzed/spicedb/pkg/datastore"
 	core "github.com/authzed/spicedb/pkg/proto/core/v1"
-	pkgruntime "github.com/authzed/spicedb/pkg/runtime"
 	"github.com/authzed/spicedb/pkg/tuple"
 )
 
@@ -51,46 +46,4 @@ func ContextualizedCaveatFrom(name string, context map[string]any) (*core.Contex
 		}
 	}
 	return caveat, nil
-}
-
-var errOverHundredPercent = errors.New("percentage greater than 100")
-
-func parsePercent(str string, freeMem uint64) (uint64, error) {
-	percent := strings.TrimSuffix(str, "%")
-	parsedPercent, err := strconv.ParseUint(percent, 10, 64)
-	if err != nil {
-		return 0, fmt.Errorf("failed to parse percentage: %w", err)
-	}
-
-	if parsedPercent > 100 {
-		return 0, errOverHundredPercent
-	}
-
-	return freeMem / 100 * parsedPercent, nil
-}
-
-// WatchBufferSize takes a string and interprets it as
-// either a percentage of memory (as a percentage of
-// 75% of free memory as measured on startup)
-// or a humanized byte string and returns the number of
-// bytes or an error if the value cannot be interpreted.
-// Returns 0 on an empty string.
-func WatchBufferSize(sizeString string) (size uint64, err error) {
-	if sizeString == "" {
-		return 0, nil
-	}
-
-	if strings.HasSuffix(sizeString, "%") {
-		size, err := parsePercent(sizeString, pkgruntime.AvailableMemory())
-		if err != nil {
-			return 0, fmt.Errorf("could not parse %s as percentage: %w", sizeString, err)
-		}
-		return size, nil
-	}
-
-	size, err = humanize.ParseBytes(sizeString)
-	if err != nil {
-		return 0, fmt.Errorf("could not parse %s as a number of bytes: %w", sizeString, err)
-	}
-	return size, nil
 }