+++ internal/datastore/proxy/proxy_test/mock.go
@@ -72,6 +72,10 @@ func (dm *MockDatastore) RevisionFromString(s string) (datastore.Revision, error
 	return args.Get(0).(datastore.Revision), args.Error(1)
 }
 
+func (dm *MockDatastore) DefaultsWatchOptions() datastore.WatchOptions {
+	return datastore.WatchOptions{}
+}
+
 func (dm *MockDatastore) Watch(_ context.Context, afterRevision datastore.Revision, _ datastore.WatchOptions) (<-chan datastore.RevisionChanges, <-chan error) {
 	args := dm.Called(afterRevision)
 	return args.Get(0).(<-chan datastore.RevisionChanges), args.Get(1).(<-chan error)