+++ internal/datastore/proxy/singleflight.go
@@ -64,6 +64,10 @@ func (p *singleflightProxy) RevisionFromString(serialized string) (datastore.Rev
 	return p.delegate.RevisionFromString(serialized)
 }
 
+func (p *singleflightProxy) DefaultsWatchOptions() datastore.WatchOptions {
+	return p.delegate.DefaultsWatchOptions()
+}
+
 func (p *singleflightProxy) Watch(ctx context.Context, afterRevision datastore.Revision, options datastore.WatchOptions) (<-chan datastore.RevisionChanges, <-chan error) {
 	return p.delegate.Watch(ctx, afterRevision, options)
 }