+++ pkg/cmd/testserver/testserver.go
@@ -111,7 +111,7 @@ func (c *Config) Complete(ctx context.Context) (RunnableTestServer, error) {
 				ExpiringRelationshipsEnabled:    true,
 				CaveatTypeSet:                   cts,
 			},
-			1*time.Second,
+			v1svc.ServerWatchConfig{},
 		)
 	}